feature: Joined Communities functional, Post stats on group; Bug fix on menu; Fully commented
The 'Communities You Joined' section is now working, rather than just dispalying a default message that you don't belong to any other communities. Adding someone to a community means they will see it show up on their end when they login, and they can add posts.

The groups list now show the amount of posts added to a group. Increment and decrement for adding or deleting posts.

A bug was present on initial login where the menu wouldn't show 'Logout'. Refereshing the page would show it though. Fixed that.

All files are now commented in detail. I spent most of the day at work getting that done, rather than listen to training.
